+++ app/node/test/node.go
@@ -19,10 +19,10 @@ import (
 	"github.com/filecoin-project/venus/fixtures/fortest"
 	"github.com/filecoin-project/venus/pkg/config"
 	"github.com/filecoin-project/venus/pkg/constants"
-	"github.com/filecoin-project/venus/pkg/util/blockstoreutil"
 	"github.com/filecoin-project/venus/pkg/util/ffiwrapper/impl"
 	"github.com/filecoin-project/venus/pkg/wallet"
 	gengen "github.com/filecoin-project/venus/tools/gengen/util"
+	blockstoreutil "github.com/filecoin-project/venus/venus-shared/blockstore"
 	"github.com/filecoin-project/venus/venus-shared/types"
 )
