+++ pkg/chain/utils.go
@@ -3,7 +3,7 @@ package chain
 import (
 	"context"
 
-	"github.com/filecoin-project/venus/pkg/util/blockstoreutil"
+	blockstoreutil "github.com/filecoin-project/venus/venus-shared/blockstore"
 	"github.com/filecoin-project/venus/venus-shared/types"
 
 	blockFormat "github.com/ipfs/go-block-format"
