+++ pkg/repo/mem.go
@@ -5,7 +5,7 @@ import (
 	"sync"
 
 	"github.com/filecoin-project/venus/pkg/repo/fskeystore"
-	"github.com/filecoin-project/venus/pkg/util/blockstoreutil"
+	blockstoreutil "github.com/filecoin-project/venus/venus-shared/blockstore"
 
 	"github.com/ipfs/go-datastore"
 	dss "github.com/ipfs/go-datastore/sync"
