+++ pkg/repo/repo.go
@@ -3,7 +3,7 @@ package repo
 import (
 	"github.com/filecoin-project/venus/pkg/config"
 	"github.com/filecoin-project/venus/pkg/repo/fskeystore"
-	"github.com/filecoin-project/venus/pkg/util/blockstoreutil"
+	blockstoreutil "github.com/filecoin-project/venus/venus-shared/blockstore"
 	"github.com/ipfs/go-datastore"
 )
