+++ venus-shared/actors/builtin_actors.go
@@ -17,8 +17,8 @@ import (
 	cbor "github.com/ipfs/go-ipld-cbor"
 	"github.com/ipld/go-car"
 
-	"github.com/filecoin-project/venus/pkg/util/blockstoreutil"
 	"github.com/filecoin-project/venus/venus-shared/actors/adt"
+	blockstoreutil "github.com/filecoin-project/venus/venus-shared/blockstore"
 )
 
 //go:embed builtin-actors-code/*.tar.zst
